requirements.txt
streamlit
yt-dlp
ffmpeg-python
openai-whisper
scikit-learn
torch
torchaudio
numpy
pandas

